@@ -1,6 +1,6 @@
 # This migration comes from blacklight (originally 20140320000000)
 # frozen_string_literal: true
-class AddPolymorphicTypeToBookmarks < ActiveRecord::Migration
+class AddPolymorphicTypeToBookmarks < ActiveRecord::Migration[4.2]
   def change
     add_column(:bookmarks, :document_type, :string) unless Bookmark.connection.column_exists? :bookmarks, :document_type
     
