@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20160328222157)
-class AddLdapAttrsToUser < ActiveRecord::Migration
+class AddLdapAttrsToUser < ActiveRecord::Migration[4.2]
   def self.up
     add_column :users, :display_name, :string
     add_column :users, :address, :string
