@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20160328222158)
-class AddAvatarsToUsers < ActiveRecord::Migration
+class AddAvatarsToUsers < ActiveRecord::Migration[4.2]
   def self.up
     add_column :users, "avatar_file_name",    :string
     add_column :users, "avatar_content_type", :string
