@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20160328222165)
-class CreateFeaturedWorks < ActiveRecord::Migration
+class CreateFeaturedWorks < ActiveRecord::Migration[4.2]
   def change
     create_table :featured_works do |t|
       t.integer :order, default: 5
