@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20160328222231)
-class CreateUserStats < ActiveRecord::Migration
+class CreateUserStats < ActiveRecord::Migration[4.2]
   def change
     create_table :user_stats do |t|
       t.integer :user_id
