@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20160328222232)
-class CreateWorkViewStats < ActiveRecord::Migration
+class CreateWorkViewStats < ActiveRecord::Migration[4.2]
   def change
     create_table :work_view_stats do |t|
       t.datetime :date
