@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20160328222233)
-class AddWorksToUserStats < ActiveRecord::Migration
+class AddWorksToUserStats < ActiveRecord::Migration[4.2]
   def self.up
     add_column :user_stats, :work_views, :integer
     add_column :work_view_stats, :user_id, :integer
