@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20160427155928)
-class CreateOperations < ActiveRecord::Migration
+class CreateOperations < ActiveRecord::Migration[4.2]
   def change
     create_table :curation_concerns_operations do |t|
       t.string :status
