@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20161012182404)
-class CreateSipityWorkflowMethods < ActiveRecord::Migration
+class CreateSipityWorkflowMethods < ActiveRecord::Migration[4.2]
   def change
     create_table :sipity_workflow_methods do |t|
       t.string   "service_name",                    null: false
