@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20161021180154)
-class CreatePermissionTemplateAccess < ActiveRecord::Migration
+class CreatePermissionTemplateAccess < ActiveRecord::Migration[4.2]
   def change
     create_table :permission_template_accesses do |t|
       t.references :permission_template, foreign_key: true
