@@ -1,5 +1,5 @@
 # This migration comes from hyrax (originally 20170131142607)
-class AddPermissionTemplateToSipityWorkflow < ActiveRecord::Migration
+class AddPermissionTemplateToSipityWorkflow < ActiveRecord::Migration[4.2]
   def change
     add_column :sipity_workflows, :permission_template_id, :integer, index: true
     remove_index :sipity_workflows, :name
