@@ -1,5 +1,5 @@
 # This migration comes from mailboxer_engine (originally 20131206080417)
-class AddMissingIndices < ActiveRecord::Migration
+class AddMissingIndices < ActiveRecord::Migration[4.2]
   def change
     # We'll explicitly specify its name, as the auto-generated name is too long and exceeds 63
     # characters limitation.
