@@ -1,6 +1,6 @@
 # This migration comes from active_fedora_noid_engine (originally 20161021203429)
 # frozen_string_literal: true
-class RenameMinterStateRandomToRand < ActiveRecord::Migration
+class RenameMinterStateRandomToRand < ActiveRecord::Migration[4.2]
   def change
     rename_column :minter_states, :random, :rand
   end
