@@ -4,7 +4,7 @@ class WorkIndexer < Hyrax::WorkIndexer
     super.tap do |solr_doc|
       # Index the ingest dates as integer timestamps so we can facet on it.
       # '_isi' suffix implies integer, stored, indexed
-      solr_doc['mdpi_timestamp_isi'] = object.mdpi_date.try(:strftime, '%Y%m%d').to_i
+      solr_doc['mdpi_timestamp_isi'] = object.mdpi_date.try(:strftime, '%Y%m%d').try(:to_i)
     end
   end
 end
