@@ -24,16 +24,27 @@ class SolrDocument
   # Recommendation: Use field names from Dublin Core
   use_extension(Blacklight::Document::DublinCore)
 
-  # Do content negotiation for AF models. 
+  # Do content negotiation for AF models.
 
   use_extension( Hydra::ContentNegotiation )
 
   def filename
-    File.basename(fetch(Solrizer.solr_name(:file_name, :stored_searchable), ['Unknown']).first)
+    @filename ||= begin
+      full_filename = fetch(Solrizer.solr_name(:file_name, :stored_searchable)).first
+      if full_filename
+        File.basename(full_filename)
+      else
+        'Unknown'
+      end
+    end
   end
 
   def file_size
     fetch(Solrizer.solr_name(:file_size, Solrizer::Descriptor.new(:long, :stored, :indexed)), [])
+  end
+
+  def file_path
+    fetch(Solrizer.solr_name(:file_path, :symbol), [])
   end
 
   def quality_level
@@ -52,14 +63,14 @@ class SolrDocument
     fetch('mdpi_timestamp_isi', [])
   end
 
-  def original_checksum
-    fetch(Solrizer.solr_name(:original_checksum, :symbol), [])
+  def md5_checksum
+    fetch(Solrizer.solr_name(:md5_checksum, :symbol), [])
   end
-  
+
   def system_create
     fetch(Solrizer.solr_name(:system_create, :stored_sortable, type: :date), [])
   end
-  
+
   def system_modified
     fetch(Solrizer.solr_name(:system_modified, :stored_sortable, type: :date), [])
   end
@@ -88,8 +99,12 @@ class SolrDocument
     fetch(Solrizer.solr_name(:file_format_long_name, :symbol), [])
   end
 
-  def codec_type
-    fetch(Solrizer.solr_name(:codec_type, :stored_searchable), [])
+  def audio_codec_type
+    fetch(Solrizer.solr_name(:audio_codec_type, :stored_searchable), [])
+  end
+
+  def video_codec_type
+    fetch(Solrizer.solr_name(:video_codec_type, :stored_searchable), [])
   end
 
   def codec_name
@@ -100,12 +115,24 @@ class SolrDocument
     fetch(Solrizer.solr_name(:codec_long_name, :stored_searchable), [])
   end
 
-  def duration
-    fetch(Solrizer.solr_name(:duration, :stored_searchable, type: :integer), [])
+  def video_width
+    fetch(Solrizer.solr_name(:video_width, :stored_searchable), [])
+  end
+
+  def video_height
+    fetch(Solrizer.solr_name(:video_height, :stored_searchable), [])
+  end
+
+  def format_duration
+    fetch(Solrizer.solr_name(:format_duration, :stored_searchable), [])
   end
 
   def bit_rate
-    fetch(Solrizer.solr_name(:bit_rate, :stored_searchable, type: :integer), [])
+    fetch(Solrizer.solr_name(:bit_rate, :stored_searchable), [])
+  end
+
+  def format_sample_rate
+    fetch(Solrizer.solr_name(:format_sample_rate, :stored_searchable), [])
   end
 
   def unit_of_origin
@@ -143,6 +170,30 @@ class SolrDocument
       recents
     end
   end
+
+  def mes_events
+    @mes_events ||= begin
+      Hyrax::Preservation::Event.search_with_conditions(hasEventRelatedObject_ssim: fetch(:id), premis_event_type_ssim: 'mes')
+    end.sort_by do |mes|
+      # Sort first by premis event date time, if found; otherwise by the system create date.
+      mes["premis_event_date_time_dtsim"] || mes['system_create_dtsi']
+    end
+  end
+
+  def current_mes_event
+    mes_events.last
+    # @current_mes_event ||= Hyrax::Preservation::Event.find(mes_events&.sort_by { |dt| dt["premis_event_date_time_dtsim"] }&.last&.id)
+  end
+
+  def previous_mes_event
+    mes_events[-2]
+    # @previous_mes_event ||= Hyrax::Preservation::Event.find(mes_events&.sort_by { |dt| dt["premis_event_date_time_dtsim"] }[-2]&.id)
+  end
+
+  def current_mes_event_changed?
+    !!( current_mes_event && previous_mes_event && (current_mes_event['premis_event_outcome_tesim'] != previous_mes_event['premis_event_outcome_tesim']) )
+  end
+
   def hardware
     fetch(Solrizer.solr_name(:hardware, :stored_searchable), [])
   end
