@@ -14,8 +14,8 @@ class User < ActiveRecord::Base
   include Blacklight::User
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable and :omniauthable
-  devise :database_authenticatable, :registerable,
-         :recoverable, :rememberable, :trackable, :validatable,
+  devise :database_authenticatable,
+         :rememberable, :trackable, :validatable,
          :omniauthable, :omniauth_providers => [:cas]
 
   # Method added by Blacklight; Blacklight uses #to_s on your
