@@ -13,6 +13,7 @@ Rails.application.routes.draw do
   devise_for :users, :controllers => { :omniauth_callbacks => "users/omniauth_callbacks" }
   mount Qa::Engine => '/authorities'
   mount Hyrax::Engine, at: '/'
+
   resources :welcome, only: 'index'
   root 'hyrax/homepage#index'
   curation_concerns_basic_routes
@@ -36,10 +37,5 @@ Rails.application.routes.draw do
   mount Hyrax::Preservation::Engine, at: '/preservation'
   concern :range_searchable, BlacklightRangeLimit::Routes::RangeSearchable.new
 
-  # Adds routes to large file interactions provided through Phydo::StorageControllerBehavior
-  get '/concern/file_sets/:id/file_status' => 'hyrax/file_sets#file_status', as: :file_status_hyrax_file_set
-  get '/concern/file_sets/:id/stage' => 'hyrax/file_sets#stage', as: :stage_hyrax_file_set
-  get '/concern/file_sets/:id/unstage' => 'hyrax/file_sets#unstage', as: :unstage_hyrax_file_set
-  get '/concern/file_sets/:id/fixity' => 'hyrax/file_sets#fixity', as: :fixity_hyrax_file_set
-
+  mount ExternalStorage::Engine, at: '/'
 end
