@@ -1,4 +1,4 @@
-class DeviseCreateUsers < ActiveRecord::Migration[5.0]
+class DeviseCreateUsers < ActiveRecord::Migration[4.2][5.0]
   def change
     create_table :users do |t|
       ## Database authenticatable
