@@ -1,4 +1,4 @@
-class CreateQaLocalAuthorities < ActiveRecord::Migration[5.0]
+class CreateQaLocalAuthorities < ActiveRecord::Migration[4.2][5.0]
   def change
     create_table :qa_local_authorities do |t|
       t.string :name
