@@ -14,17 +14,31 @@ RSpec.describe "WGBH ingest example 3 batch" do
     expect(FileSet.all.count).to eq @sip_paths.count
   end
 
-  it 'ingests a single preservation event with PREMIS event type of "Ingest" and PREMIS agent of mailto:admin@example.org', :clean_fedora, :large_ingest do
+
+  # TODO: Lots of expectations happening in this example. Better way?
+  it 'ingests 2 preservation events: one for ingest, and one for a fixity check, both having PREMIS agent as mailto:admin@example.org', :clean_fedora, :large_ingest do
     FileSet.all.each do |file_set|
-      # Expect exactly 1 preservation event.
-      expect(file_set.preservation_events.count).to eq 1
+      # Expect exactly 2 preservation events.
+      expect(file_set.preservation_events.count).to eq 2
 
-      # Expect the PREMIS event type to be that which was specified in the ingest config YAML.
-      premis_event_type_uri = file_set.preservation_events.first.premis_event_type.first.to_uri.to_s
-      expect(premis_event_type_uri).to eq 'http://id.loc.gov/vocabulary/preservation/eventType/ing'
+      # Grab all events that represent a fixity check (there should noly be 1 per FileSet).
+      fixity_check_events = file_set.preservation_events.select do |event|
+        event.premis_event_type.first.to_uri.to_s == 'http://id.loc.gov/vocabulary/preservation/eventType/fix'
+      end
 
-      # Expect the PREMIS agent to be that which was specified in the ingest config YAML.
-      premis_agent = file_set.preservation_events.first.premis_agent.first.to_uri.to_s
+      # Expect exactly 1 fixity check event.
+      expect(fixity_check_events.count).to eq 1
+
+      # Grab all events that represent the ingest (there should noly be 1 per FileSet).
+      ingest_events = file_set.preservation_events.select do |event|
+        event.premis_event_type.first.to_uri.to_s == 'http://id.loc.gov/vocabulary/preservation/eventType/ing'
+      end
+
+      # Expect exactly 1 ingest event.
+      expect(ingest_events.count).to eq 1
+
+      # Expect the PREMIS agent of the ingest event to be that which is specified in the ingest config.
+      premis_agent = ingest_events.first.premis_agent.first.to_uri.to_s
       expect(premis_agent).to eq 'mailto:admin@example.org'
     end
   end
