@@ -8,14 +8,15 @@ describe Concerns::FileSetBehavior do
   end
 
   describe 'adds properties' do
-    { bit_rate: RDF::Vocab::EBUCore.bitRate,
+    { audio_codec_type: RDF::Vocab::EBUCore.hasAudioFormat,
+      bit_rate: RDF::Vocab::EBUCore.bitRate,
       codec_long_name: RDF::Vocab::EBUCore.codecName,
       codec_name: RDF::Vocab::EBUCore.hasCodec,
-      codec_type: RDF::Vocab::EBUCore.hasMedium,
       date_generated: RDF::Vocab::EBUCore.dateCreated,
       file_format_long_name: RDF::Vocab::PREMIS.hasFormatName,
       file_format: RDF::Vocab::EBUCore.hasFileFormat,
       file_name: RDF::Vocab::EBUCore.filename,
+      file_path: RDF::Vocab::EBUCore.locator,
       format_duration: RDF::Vocab::EBUCore.duration,
       format_file_size: RDF::Vocab::EBUCore.fileSize,
       format_sample_rate: RDF::Vocab::EBUCore.sampleRate,
@@ -25,6 +26,7 @@ describe Concerns::FileSetBehavior do
       quality_level: RDF::Vocab::EBUCore.encodingLevel,
       title: RDF::Vocab::EBUCore.title,
       unit_of_origin: RDF::Vocab::EBUCore.comments,
+      video_codec_type: RDF::Vocab::EBUCore.hasVideoFormat,
       video_height: RDF::Vocab::EBUCore.height,
       video_width: RDF::Vocab::EBUCore.width }.each do |attribute, predicate|
       specify "#{attribute} => #{predicate}" do
